pyproject.toml
[project]
name = "raspberry-pi-car"
version = "0.1.0"
description = "Flask server for controlling a Raspberry Pi robot car"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "RPi.GPIO>=0.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
